pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "WAPP_Tools"
license = {text = "MIT License"}
dependencies = [
    "Pillow",
    "crc32c",
    "jsonschema",
]
version = "0.5.0"

[project.scripts]
wapp_image = "wapp_tools.wapp_image:main"
wapp = "wapp_tools.wapp:main"

[tool.setuptools.package-dir]
wapp_tools = "tools"

